requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
openai==1.51.0

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]

# File parsing
pypdf==5.0.1
python-docx==1.1.2
python-multipart
pdfplumber
docx2txt
